Time: 27 ms (68.80%) | Memory: 44.3 MB (32.26%) - LeetSync

+++ 115-distinct-subsequences/distinct-subsequences.cpp
@@ -26,6 +26,6 @@ public:
             }
         }
 
-        return dp[m][n];
+        return (int)dp[m][n];
     }
 };
